validação de campos, required

+++ app/models.py
@@ -4,6 +4,6 @@ from django.db import models
 # Create your models here.
 class Cursos(models.Model):
     cursos = models.CharField(max_length=150)
-    horas = models.IntegerField()
+    horas = models.BigIntegerField()
